% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_webex_to_bookdown.R
\name{add_webex_to_bookdown}
\alias{add_webex_to_bookdown}
\title{Add webex to bookdown}
\usage{
add_webex_to_bookdown(
  bookdown_dir = ".",
  include_dir = "include",
  script_dir = "R",
  output_format = "bookdown::bs4_book"
)
}
\arguments{
\item{bookdown_dir}{The base directory for your bookdown project}

\item{include_dir}{The directory where you want to put the css and js files (defaults to "include")}

\item{script_dir}{The directory where you want to put the .R script (defaults to "R")}

\item{output_format}{The bookdown format you want to add webex to (defaults to "bookdown::bs4_book") This is typically your default HTML format in the _output.yml file}
}
\description{
Add webex to bookdown
}
