% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_webex_to_bookdown.R
\name{add_webexercises_to_bookdown}
\alias{add_webexercises_to_bookdown}
\title{Add webexercises helper files to bookdown}
\usage{
add_webexercises_to_bookdown(
  bookdown_dir = ".",
  include_dir = "include",
  script_dir = "R",
  output_format = c("bs4_book", "gitbook", "html_book", "tufte_html_book"),
  render = FALSE
)
}
\arguments{
\item{bookdown_dir}{The base directory for your bookdown project}

\item{include_dir}{The directory where you want to put the css and
js files (defaults to "include")}

\item{script_dir}{The directory where you want to put the .R script
(defaults to "R")}

\item{output_format}{The bookdown format you want to add
webexercises to (defaults to "bs4_book") This is typically your
default HTML format in the _output.yml file.}

\item{render}{Whether to render the book after updating (defaults to FALSE).}
}
\description{
Adds the necessary helper files to an existing bookdown project and
edits the _output.yml and _bookdown.yml files accordingly. If the
directory does not have a bookdown project in it, a template
project will be set up.
}
