% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{checkbox}
\alias{checkbox}
\title{Create a set of checkboxes}
\usage{
checkbox(..., name = NULL)
}
\arguments{
\item{...}{questions and answers in the format of "question text" = TRUE}

\item{name}{The name to record if you are saving your answers with a script (a single name or list of names)}
}
\value{
a string with HTML for the checkbox group
}
\description{
Create a set of checkboxes
}
