% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{eng_hint}
\alias{eng_hint}
\title{Create an engine to allow block-style hints: like ```{hint} button text / hint text
The first line of the block is used as the button text, all other lines are rendered 
as markdown and included in the hidden-by-default div}
\usage{
eng_hint(options)
}
\description{
Create an engine to allow block-style hints: like ```{hint} button text / hint text
The first line of the block is used as the button text, all other lines are rendered 
as markdown and included in the hidden-by-default div
}
