% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape a string for regex}
\usage{
escape_regex(string)
}
\arguments{
\item{string}{A string to escape}
}
\value{
A string
}
\description{
Escape a string for regex
}
\examples{
escape_regex("library(tidyverse)")
}
