% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{fitb}
\alias{fitb}
\title{Create fill-in-the-blank question}
\usage{
fitb(answer, width = calculated_width, num = FALSE, tol = NULL,
  calculator = num, calculator_digits = 4, ignore_case = FALSE,
  ignore_ws = TRUE, regex = FALSE)
}
\arguments{
\item{answer}{The correct answer}

\item{width}{Width of the input box in characters. Defaults to the length of the longest answer.}

\item{num}{whether the input is numeric, in which case allow for leading zeroes to be omitted}

\item{tol}{the tolerance within which numeric answers will be accepts; i.e. (response - true.answer) < tol = a correct response. Implies num=TRUE}

\item{calculator}{whether the input should be eval'd in the local javascript environment. Defaults to `num``}

\item{calculator_digits}{number significant digits to show}

\item{ignore_case}{Whether to ignore case (capitalization)}

\item{ignore_ws}{Whether to ignore white space}

\item{regex}{Whether to use regex to match answers (concatenates all answers with `|` before matching)}
}
\description{
Create fill-in-the-blank question
}
\details{
Writes html code that creates an input box widget. Call this function inline in an RMarkdown document. See the Web Exercises RMarkdown template for examples.
}
