% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_forms.R
\name{form_start}
\alias{form_start}
\title{Start a form}
\usage{
form_start(url = "")
}
\arguments{
\item{url}{The URL to handle the POST request with the form items, leave blank to just check answers}
}
\value{
a string with the html to start the form
}
\description{
Start a form
}
