% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_clean.R
\name{html_clean}
\alias{html_clean}
\title{Knitr function to convert to an HTML document}
\usage{
html_clean(..., lightbox = FALSE, thumbnails = FALSE,
  gallery = FALSE)
}
\description{
Format for converting from R Markdown to an HTML document including classroom exercises.
Basically just adds dependencies to rmdformats::html_clean
}
