% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{longmcq}
\alias{longmcq}
\title{Longer MCQs with Radio Buttons}
\usage{
longmcq(opts)
}
\arguments{
\item{opts}{Vector of alternatives. The correct answer is the
element(s) of this vector named 'answer'.}
}
\description{
Longer MCQs with Radio Buttons
}
\details{
Writes html code that creates a radio button widget, with a
  single correct answer. This is more suitable for longer answers. Call this function inline in an RMarkdown
  document. See the Web Exercises RMarkdown template for further
  examples.
}
\examples{
# What is a p-value?
opts <- c(
  "the probability that the null hypothesis is true",
  answer = paste("the probability of the observed, or more extreme, data",
                 "under the assumption that the null-hypothesis is true"),
  "the probability of making an error in your conclusion"
)
longmcq(opts)

}
