% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{mcq}
\alias{mcq}
\title{Create multiple choice question}
\usage{
mcq(opts)
}
\arguments{
\item{opts}{Vector of alternatives. The correct answer is the element(s) of this vector named 'answer'. See the Web Exercises RMarkdown template for examples.}
}
\description{
Create multiple choice question
}
\details{
Writes html code that creates an option box widget. Call this function inline in an RMarkdown document. See the Web Exercises RMarkdown template for examples.
}
