% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{round2}
\alias{round2}
\title{Round up from .5}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{}

\item{digits}{integer indicating the number of decimal places (‘round’) or significant digits (‘signif’) to be used.}
}
\description{
Round up from .5
}
\details{
Implements rounding using the "round up from .5" rule, which is more conventional than the "round to even" rule implemented by R's built-in \code{\link{round}} function. This implementation was taken from \link{https://stackoverflow.com/a/12688836}.
}

