% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{strip_lzero}
\alias{strip_lzero}
\title{Strip leading zero from numeric string}
\usage{
strip_lzero(x)
}
\arguments{
\item{x}{a numeric string (or number that can be converted to a string)}
}
\value{
a string with leading zero removed
}
\description{
Strip leading zero from numeric string
}
