% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{style_widgets}
\alias{style_widgets}
\title{Change webex widget style}
\usage{
style_widgets(default = "blue", correct = "red")
}
\arguments{
\item{default}{the colour of the widgets when the correct answer is not filled in (defaults to blue)}

\item{correct}{the colour of the widgets when the correct answer not filled in (defaults to red)}
}
\description{
Change webex widget style
}
\details{
Call this function inline in an RMarkdown document to change the default and correct colours using any valid HTML colour word (e.g., red, rgb(255,0,0), hsl(0, 100%, 50%) or #FF0000).
}
