% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{total_correct}
\alias{total_correct}
\title{Display total correct}
\usage{
total_correct(elem = "span", args = "")
}
\arguments{
\item{elem}{The html element to display (e.g., div, h3, p, span)}

\item{args}{Optional arguments for css classes or styles}
}
\value{
A string with the html for displaying a total correct element
}
\description{
Display total correct
}
\examples{
total_correct()     # <div  id="total_correct"></div>
total_correct("h3") # <h3  id="total_correct"></h3>
total_correct("p", "style='color: red;'")
total_correct("div", "class='customclass'")
}
