% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{unhide}
\alias{unhide}
\title{End hidden HTML content}
\usage{
unhide()
}
\description{
End hidden HTML content
}
\details{
Call this function inline in an RMarkdown document to mark the end of hidden content (see the Web Exercises RMarkdown Template for examples).
}
\examples{
# just produce the closing </div> 
unhide()
}
\seealso{
\code{hide}
}
