% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{fitb}
\alias{fitb}
\title{Create fill-in-the-blank question}
\usage{
fitb(answer, width = 3, num = FALSE, ignore_case = FALSE,
  ignore_ws = TRUE)
}
\arguments{
\item{answer}{The correct answer}

\item{width}{Width of the input box in characters}

\item{num}{whether the input is numeric, in which case allow for leading zeroes to be omitted}

\item{ignore_case}{Whether to ignore case (capitalization)}

\item{ignore_ws}{Whether to ignore white space}
}
\description{
Create fill-in-the-blank question
}
\details{
Writes html code that creates an input box widget. Call this function inline in an RMarkdown document. See the Web Exercises RMarkdown template for examples.
}
