% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/webex_fns.R
\name{fitb}
\alias{fitb}
\title{Create fill-in-the-blank question}
\usage{
fitb(answer, size = 3, ignore_case = FALSE, ignore_ws = TRUE)
}
\arguments{
\item{answer}{The correct answer}

\item{size}{Width of the input box in characters}

\item{ignore_case}{Whether to ignore case (capitalization)}

\item{ignore_ws}{Whether to ignore white space}
}
\description{
Create fill-in-the-blank question
}
\details{
Writes html code that creates an input box widget. Call this function inline in an RMarkdown document. See the Web Exercises RMarkdown template for examples.
}

