% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_default.R
\name{webex_default}
\alias{webex_default}
\title{Create default webex document}
\usage{
webex_default(...)
}
\arguments{
\item{...}{Additional function arguments to pass to html_document}
}
\description{
This function wraps \code{rmarkdown::html_document} to configure compilation to embed the default webex CSS and JavaScript files in the resulting HTML.
}
